refactor : code splitted article comment layouts
